Add files via upload
Snake with 8 LEDs and your Microcontroller ATMEL 89C52. This is, turning all the LEDs ON one by one from the right to the left, turning them OFF in the same direction, turning them ON in the opposite direction and then OFF again in the same direction (left to right). Using your imagination, you’ll see a snake going from one side to the other; first leaving the “LEDs area”, coming back to it form the same side it left and then doing this cyclically until you disconnect the power source. Make sure you use an 11.059 MHZ Crystal if you use exactly the same code.